moved to different package so code coverage works out of the box
