+++ test/kru/collections/AbstractListTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package kru.collections.testing;
+package kru.collections;
 
 import org.junit.Test;
