+++ test/kru/collections/LinkedListTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package kru.collections.testing;
+package kru.collections;
 
 import java.util.List;
